refactor(architecture): add services folder and organize components of the admin
